Removed line manager.
